+++ .github/workflows/release.yml
@@ -12,16 +12,22 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v3
       - name: Install Go
-        uses: actions/setup-go@v3
+        uses: actions/setup-go@v4
         with:
-          go-version: '>=1.18.0'
+          go-version: '>=1.19.0'
       - name: Build binaries
         run: |
-          CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o "xspreak-$(git describe --tags)-linux-amd64"
-          CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build -o "xspreak-$(git describe --tags)-darwin-amd64"
-          CGO_ENABLED=0 GOOS=windows GOARCH=amd64 go build -o "xspreak-$(git describe --tags)-windows-amd64.exe"
+          CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build \
+            -ldflags="-X 'github.com/vorlif/xspreak/commands.Version=$(git describe --tags)'" \
+            -o "xspreak-$(git describe --tags)-linux-amd64"
+          CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build \
+            -ldflags="-X 'github.com/vorlif/xspreak/commands.Version=$(git describe --tags)'" \
+            -o "xspreak-$(git describe --tags)-darwin-amd64"
+          CGO_ENABLED=0 GOOS=windows GOARCH=amd64 go build \
+            -ldflags="-X 'github.com/vorlif/xspreak/commands.Version=$(git describe --tags)'" \
+            -o "xspreak-$(git describe --tags)-windows-amd64.exe"
       - name: Upload release artifacts
-        uses: actions/github-script@v3
+        uses: actions/github-script@v7
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           script: |
